backend/simple_services/reservation/requirements.txt
fastapi==0.110.0
pydantic==2.6.4
uvicorn>=0.15.0,<0.16.
mysql-connector-python==8.3.0
pymysql==1.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.28

